refactor(runtime): decouple/tidy up `useResolvedProps`
pass `stylesheetFactory` as an argument, switch to named args,
move `useControlInstances` to a separate file
