+++ packages/runtime/src/runtimes/react/stylesheet-factory.ts
@@ -1,0 +1,6 @@
+import { type Stylesheet } from '@makeswift/controls'
+
+export type StylesheetFactory = {
+  get(propName: string): Stylesheet
+  useDefinedStyles(): void
+}
